requirements.txt
fastapi
uvicorn
pydantic
python-multipart
langchain
openai
faiss-cpu
PyMuPDF
httpx
python-dotenv
